좋아요 백엔드 로직 수정(로그인,로그아웃 시 발생할수있는 로직을 수정)

+++ src/main/java/com/app/controller/review/LikeController.java
@@ -24,10 +24,13 @@ public class LikeController {
 	
 	private final LikeService likeService;
 	
-	// 게시글 좋아요 활성화
+	// 게시글 좋아요 활성화 및 토글
 	@PostMapping(consumes = "application/json")
 	public ResponseResult<?> insert(@RequestBody LikeRequestDTO likeRequestDTO, HttpSession session) throws Exception {
 		MemberDTO loginUser = (MemberDTO)session.getAttribute("loginUser");
+		if (loginUser == null) {
+			return ResponseResult.fail("로그인이 필요한 서비스입니다.");
+		}
 		likeRequestDTO.setUserId(loginUser.getUserId());
 		likeService.insert(likeRequestDTO);
 		return ResponseResult.success(null);
@@ -37,6 +40,9 @@ public class LikeController {
 	@DeleteMapping(consumes = "application/json")
 	public ResponseResult<?> delete(@RequestBody LikeRequestDTO likeRequestDTO, HttpSession session) throws Exception {
 		MemberDTO loginUser = (MemberDTO)session.getAttribute("loginUser");
+		if (loginUser == null) {
+			return ResponseResult.fail("로그인이 필요한 서비스입니다.");
+		}
 		likeRequestDTO.setUserId(loginUser.getUserId());
 		likeService.delete(likeRequestDTO);
 		return ResponseResult.success(null);
